Move BobState to a private submodule
This does not only organise things better, but most importantly the
BobStateHandle is now not in the same module as its users.  This means
it can guarantee safety about how it is initialised and use
unreachable!() to simplify it's API.
